+++ fragnet-search/src/main/java/org/squonk/fragnet/search/queries/v2/SuppliersQuery.java
@@ -1,3 +1,18 @@
+/*
+ * Copyright (c) 2019 Informatics Matters Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package org.squonk.fragnet.search.queries.v2;
 
 import org.neo4j.driver.v1.Session;
@@ -48,14 +63,6 @@ public class SuppliersQuery {
             return results;
         });
         return suppliers;
-    }
-
-    public static final void main(String[] args) {
-        GraphDB db = new GraphDB();
-        Session session = db.getSession();
-        SuppliersQuery query = new SuppliersQuery(session);
-        List<Map<String,String>> suppliers = query.getSuppliers();
-        LOG.info("Found " + suppliers.size() + " suppliers");
     }
 
 }