+++ fragnet-search/src/test/groovy/org/squonk/fragnet/search/queries/v2/SuppliersQuerySpec.groovy
@@ -1,0 +1,37 @@
+/*
+ * Copyright (c) 2019 Informatics Matters Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.squonk.fragnet.search.queries.v2
+
+import org.neo4j.driver.v1.Session
+import org.squonk.fragnet.search.model.v2.Availability
+import org.squonk.fragnet.service.GraphDB
+import spock.lang.Specification
+
+class SuppliersQuerySpec extends Specification {
+
+    void "simple search"() {
+        GraphDB db = new GraphDB()
+        Session session = db.getSession()
+        SuppliersQuery query = new SuppliersQuery(session)
+
+        when:
+        def suppliers = query.getSuppliers()
+
+        then:
+        suppliers.size() > 0
+    }
+
+}